requirements.txt
streamlit>=1.24.0
opencv-python>=4.7.0
torch>=2.0.0
ultralytics>=8.0.0
pillow>=9.0.0
numpy>=1.24.0
pytesseract>=0.3.10
face-recognition>=1.3.0
plotly>=5.13.0
pandas>=1.5.0
dlib>=19.24.0 